Various fixes:
bot.js was not using asyncronous callbacks, expecting syncronous return 
values! *TODO: Tidy up...*

Tidied up the SQL code, changed the "user" field to be uid so as to 
abandon the need for quotes...

db\repos\socres.js
+ Added function for checking if the db table exists.

pg-monitor was not specified as a dependancy in the package.json!

db\index.js
- fixed incorrect member name

db\repos\scores.js:
- Fixed extend columnset errors