Changes to expand the database into a multi-table relational database.
Two new tables, USERS and GUILDS, that store the discord guild/user id.  
These fields are unique, and a reference to them is now stored in the 
scores table instead of the actual id.

This will allow for future expansion, such as saving user and guild 
preferences.

Added the basic SQL files; still to come is the users.js and guilds.js 
class to allow MrData to run.

Added an upgrade script to upgrade the live database from 1.0.2 to this 
new version.