Initial commit.
There is an issue with systemctl not accessing/setting up the 
environment variables correctly. So this solution has been implemented.

This file will store the bot's discord connection token and its database 
username and password.

This blank version should be the only commit of this file.  Any future 
updates should ensure that this template remains blank.

This file will should be added to .gitignore so that the actual run-time 
bot token and live DB url/password are not shared.