Added ability to vary the length of the ranking table output - instead of a hardcoded 10, it can now vary between 1 and 25 (embedded messages are limited to a maximum of 25 rows).
Added a new helper to the library - containing one new function - 
parseMaxRankCount() - which takes args as a parameter, and returns the 
rank count to be used. If the command to vary the length was found, the 
args.length is reset to 0, so that the command is consumed, and the 
table will therefore be displayed.