Amends the collection of message helper functions, to allow for the asynchronous sending and editing of messages.
Adds helper functions to parse discord names (removing @ character) and id numbers, which may also be preceeded by an @ character.  The parseNumber function uses a regex, and concatenates all number-substrings into one return value.

Updates the addroletorole command so that the success or failure of the command is feed back to the user, fixing issue #25.