Preperation for mobile push notifications, addressing issue #27.
This change adds a schema (user_guild_settings) and a selection of SQL queries to create, and manipulate these user settings.

Table: user_guild_settings.
u_id, g_id - primary key, these reference the same id from users and guilds tables.
settings - jsonb - this stores the settings as a json object, which ensures that the settings will be future proof!

The table stores user settings, where the key is a compound key, using the same integer values as the users/guilds tables. This allows user settings to be stored on a per-server basis.

Using a JSON object datatype means that we can add extra settings without needing database schema changes. Future proof!

This is the first commit of a multi-part commit.