This change adds a schema (user_global_settings) and a selection of SQL queries to create, and manipulate these user settings.
Table: user_global_settings.
  u_id - primary key, references the same id from users table.
  settings - jsonb - this stores the settings as a json object - future proof!

This table will contain settings that will apply to the user globally, across all discord servers.

The table stores user settings, using the same integer the users table. The queries to manipulate the data all refer to the users table, so that data lookups can be made using the user discord id: the discord id is mapped to the primary key via the users table, and so internally we can save the settings using the primary key from the users table.

Using a JSON object datatype for the settings means that we can add extra settings fields without needing database schema changes. Future proof!

This is the second commit of a multi-part commit.