This change adds a schema (guild_settings) and a selection of SQL queries to create, and manipulate these guild settings.
Table: guild_settings.
   u_id - primary key, references the same id from guilds table.
   settings - jsonb - this stores the settings as a json object - future proof!

The table stores guild settings, using the same integer the guilds table. The queries to manipulate the data all refer to the guilds table, so that data lookups can be made using the guild discord id: the discord id is mapped to the primary key via the guilds table, and so internally we can save the settings using the primary key from the guilds table.

Using a JSON object datatype for the settings means that we can add extra settings fields without needing database schema changes. Future proof!

This is the third commit of a multi-part commit.