More changes for issue #27 - this change adds the repository layer for the new settings data.
These repositories encapsulate the logic for accessing & manipulating the settings tables.  This level of abstraction hides the data layer and provides an interface for which to access.

Three new classes have been added, one for each settings table.  The repositories provide an interface to add, update, insert, find and delete data from the tables.  The interfaces all take a discord id (user or guild) and map these to the relevant user/guild entry within the existing schema, linking our tables and hiding the logic from the user, as it does not need to be exposed. For example: the user wants settings for a discord server with discord id X - they are not concerned with having to look up the internal integer key, and using this to find the settings.