Issue #27 development. Modify the /subscribe end-point for saving settings to the database.
Modify the SQL for insert, so that on a failuer, we do nothing - this allows the code to handle the null record set.

Add a new library to acccess the settings and add push tokens into the object.  Also a new function that returns an new & empty settings object, so that it is always created in one place and additional fields to the settings object can be easily made.