This fixes issue #30.  There is a null reference error when calling library.System.getPrefix(guildId), since the message.guild object is null, and therefore accessing the id field raises an exception.
Since the custom command prefix has been implemented, users may not be using the default prefix.  Certain commands can be entered using direct messages - but since direct messages are not guild specific, the user has to use the default command prefix and not any user defined ones.

This change prevents the null reference error and additionally, if the message is a DM but the command prefix was not recognised, MrData returns a message to the user explaining they should use the default prefix, and reminds them of it.

(cherry picked from commit 91735cd4430a942dd52b9a264678e9e82c6cc3e0)