Adds new functionality for searching for discord members from a guild object.
New private function, findDiscordMember, wraps the guild.members.find() method.  This takes the parameter type and value, and the guild to search.

getDiscordMember(ident, guild) is now a wrapper for two new functions, and invokes the correct function based on the content of the ident parameter.

getDiscordMemberByName implements the functionality that was previously performed in getDiscordMember.

getDiscordMemberById attempts to find the member using the id number.

Both of the new functions call the private function, findDiscordMember, passing "displayName" or "id" respectively.