Changes to address issue #26, more settings access functions and some general tidying.
Settings are now cached in the discord client upon loading of the bot. During the caching,
the settings JSON will be checked against the version number of the current settings template. If the settings is old, it will be upgraded using a new upgrade function.

Plug n play of the command modules begins with this commit, moving the initialisation of the commands out of bot.js and into a new library, Commands.

system.js:
New functions to access settings parameters programatically, in the form of the following:
+ saveParameter(guildId, paramName, paramValue, client)
This will access the field: settings['paramName'] = paramValue, meaning we don't need to use dot notation which requires knowledge of the parameter name during coding. This logic means we can add new fields without needing to add a new get/set access function every time we add a new parameter.

Guild Settings has additional fields, including "modified", which will be used to determine whether the setting should be committed to the database.  Currently we are committing to the database every time a setting changes.  We should be able to move away from this manner of saving, now that settings are cached in the discord object.  It should be possible to make periodic checks and commit the changes to reduce database accesses. Additionally, changes could be saved during shut down of the bot.

settings.js:
New function to upgrade a guild settings json, using the npm module, compare-versions.  The version number of the settings is now stored within the settings JSON, using the package version of the bot - Helper.packageVersion() function.

Bot.js:
New code to load and cache the settings of every guild during startup.

The library module is now stored in global.library - this finally solves the circular referencing that sometimes occurs when one library module, such as Helper, requires the functions in Format, when Format already refers to Helper it causes runtime exceptions.  And requiring '../../library' has never worked.

Now library modules can refer to other library modules using global.library.Modulename.XYZ, without error.