Urgent fix for Issue #31: The default variable assignments for JSON and other objects is not to pass by value, but by reference.  This means that changing the value of the copy modifies the original. This change modifies the assignment of the settings json to protect contents.
Various other changes, such as modifying the help text templated parameters functions, for issue #24, so that they are not so specific to help text description, but so they can be used more generally to replace string templates with a dynamic value, regardless of the context.

+ settings.js
Added two new functions that return a descriptive settings json, describing each field.

+ format.js
Added a new library function to return the type of an object in a better way than the default typeof operator.

+ discord.js
Added a markdown property, which contains the discord mark down

+config.js
Changed the assignment operations for objects.  Added new text template, @SERVERNAME.
New function, replaceJsonTextParameters, loop through and replace template strings in a json object, replacing any template occuring in the name or in the value. Switch statement, using the typeof function, allows different functionality depending on the value of the field. Objects will be treated differently, since the default behaviour is to pass by reference and we don't want this.

+ library/index.js
New jsdoc header for the collateArgs function