Further implementation of guild settings caching. Issue #21 requests user configuration - this change improves caching of the guild configurations, in order to minimise the number of database accesses.
Headers updated for issue #22. Ensuring objects are passed by value where required, for issue #31.

+ System/system.js:
new functions (cacheSettings, cacheParameter) allow a given property of the settings json (we are calling them parameters, although this might be confusing - consider renaming), or the whole object, to be cached into the client.guildSettings for a given guild id. Exceptions will be thrown where the guild id or the client parameters are null.

Re-factor saveParameter, savePrefix, to use the new caching functions where appropriate.

Add missing JSDoc headers for the other exported functions.

saveGuildSettings now throws exceptions.

Renamed getGuildSettings to be more meaningful -> getGuildSettingsFromDb. And added a new function, also named getGuildSettings, which wraps getGuildSettingsFromDb to preserve the existing logic. This new function attempts to acccess the cached settings before accessing the Db. The client parameter has a default value of null to preserve existing logic.

+ bot.js:
New and old version numbers are now logged when upgrading settings objects during the client.on.ready handler.

Removed a re-declaration of a for loop index variable that was already declared.

+ commands/datahelp.js:
Execute is now async, and the function will wait for the database to return the command prefix, in order to fix issue #24 dynamic prefixes.