New functionality to add a request for funding to the "please wait..." messages
+ helper.js:
Modify the createEmbed function to add an image, a thumbnail, and a field seeking funding.  The thumbnail will be a square to the right of the embed text, and the image will be a banner along the bottom of the embed.  The text will be generated using a new function (see below) from a given array of strings.

Two constants have been added, fundingTitles, and fundingTexts, which are used to randomly generate a request for funding whenever the createEmbed is used.

This will be monitored over the coming days, so as not to flood the discord channels with begging texts. Perhaps add a throttle so not every message contains the request.

+ format.js:
Add a function to generate random numbers between two values. Add a function to pick a random string from an array of strings.

TODO: Consider re-factoring the format library to find a more logical place for these two functions