Add new library, funding, and moved the titles and texts from helper into the new library.  Incremented version number.
+ System.js:
New function, getCachedParameter - gets a property from the cached copy of the settings json.

+ Helper.js:
Removed the titles and texts which are now in the funding library.

createEmbed() now has some basic throttling, based on the config.timeBetweenAdverts property (whose unit is seconds). When we display an advert, a flag is set in the settings to indicate that no more adverts should be shown - a timer is raised which will reset this flag at the given interval.

The function is also used when sending or editting a message.  If the original message contains an advert, we don't want to remove it - so the function now passes and checks the message ID against a cached one in the settings json, and if they match it adds the advert back.

Add a footer with icon to the message, outputting the package name and version.

+ Settings.js
Fixed a typo. Added the new properties to the newSettings json.

+ Format.js:
New functions to encode and decode a 3-byte data field into a colour - since an RGB format colour is a 32-bit int, there are 3 spare bytes when storing a 64-bit number.

+ Config.js
New function to return the avatar of the bot from the client object.

Removed unnecessary function prototypes

+ commands/systeminformation.js

Removed unnecessary require statements.  Fixed an undeclared variable.