Fixes the funding throttling, which was committed ysterday but was not working, as when editting a message the stored ID and the actual ID would always be the same, so the advert = false was being ignored.
Also, when the ready message was received in bot.js, the cache was being overwritten without checking whether it was modified, because the client was not being passed!

System.js:
Add jsdoc header to getCachedParameter

Helper.js:
Modified the logic in createEmbed so that messages are not given an advert when showAdvert is false - the exception is if we are editting a message with an advert in it - now a unique id is checked, which is only set when an advert is attached during the original send.

Systeminformation.js:
Added the moment-timezone require, which was wrongly removed recently.

bot.js:
Pass the client into the call to getGuildSettings()