creator.js & mrdata.js: Moved the existing creator function into new command, iscreator.js.
These new commands use the new userCard function to display the user information of the bot creator (creator), and the bot itself (mrdata)

helper.js:
Modify the rich embed function to allow the passing of all the parameters used in a discord embed. These have default values so as not to break existing functionality.

Two new functions to display discord user data as a "user card" in an embed message.