A possible race condition exists, where the clientignoreMyself flag as been set to false, whilst at the same time the bot is sending a DM to someone (for example, responding to the datahelp command, or sending a message because a guild has been created/deleted).
When a message is received, if the message was sent by MrData, client.ignoreMyself is checked and if true, we exit without performing the rest of the event handler. This change adds a check of whether the message was a DM and, if so, we will now always ignore a DM regardless of the state of the ignoreMyself flag.